# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LiberFly Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/index": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get All Books",
                "operationId": "9988f960b568df1031bfcda7382c909e",
                "responses": {
                    "200": {
                        "description": "Returns user information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "59"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "email@email.com"
                                        },
                                        "edition": {
                                            "type": "string",
                                            "example": 29
                                        },
                                        "publishing": {
                                            "type": "string",
                                            "example": "739.073.280-60"
                                        },
                                        "user": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro interno no servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/show": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get Books information",
                "operationId": "5a28635064a124e432f129251c828936",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of the user to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "59"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "email@email.com"
                                        },
                                        "edition": {
                                            "type": "string",
                                            "example": 29
                                        },
                                        "publishing": {
                                            "type": "string",
                                            "example": "739.073.280-60"
                                        },
                                        "user": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro interno no servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/store": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Armazenar um novo livro",
                "description": "Armazena um novo livro no sistema.",
                "operationId": "4f6ddb28b9f8f2143526c16ff6f30494",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Quimica"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Jony"
                                    },
                                    "edition": {
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "publishing": {
                                        "type": "string",
                                        "example": "Minas Book"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "livro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "As Aventuras de PII"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "Nicolau Quinto"
                                        },
                                        "edition": {
                                            "type": "string",
                                            "example": "Segunda"
                                        },
                                        "publishing": {
                                            "type": "string",
                                            "example": "Ed Sentopeia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de Validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "{name:required, author:required, edition:required, publishing:required}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/books/delete": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Deletar um livro",
                "description": "Deletar um livro no sistema.",
                "operationId": "2c019ddb4cc9e70a55d0fa20f913634b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Authorization Token not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Book not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}